requirements.txt
aiohttp
pillow
flake8
numpy
scipy
